mokel

+++ database/migrations/2026_02_11_141016_create_perizinans_table.php
@@ -6,25 +6,33 @@ use Illuminate\Support\Facades\Schema;
 
 return new class extends Migration
 {
-    /**
-     * Run the migrations.
-     */
     public function up(): void
     {
-        Schema::table('izins', function (Blueprint $table) {
-            $table->enum('status', ['menunggu', 'disetujui', 'ditolak'])
-                  ->default('menunggu')
-                  ->after('tanggal_izin');
-        });
+        if (!Schema::hasColumn('izins', 'status')) {
+
+            Schema::table('izins', function (Blueprint $table) {
+
+                $table->enum('status', [
+                    'menunggu',
+                    'disetujui',
+                    'ditolak'
+                ])->default('menunggu');
+
+            });
+
+        }
     }
 
-    /**
-     * Reverse the migrations.
-     */
     public function down(): void
     {
-       Schema::table('izins', function (Blueprint $table) {
-            $table->dropColumn('status');
-        });
+        if (Schema::hasColumn('izins', 'status')) {
+
+            Schema::table('izins', function (Blueprint $table) {
+
+                $table->dropColumn('status');
+
+            });
+
+        }
     }
 };
